removed list, combined with show
